# src/backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Clinic Management API","description":"REST API for managing patients and appointments","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Patients","description":"Manage clinic patients"},{"name":"Appointments","description":"Manage clinic appointments"}],"paths":{"/api/patients/{id}":{"get":{"tags":["Patients"],"summary":"Get patient by ID","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PatientResponseDTO"}}}}}},"put":{"tags":["Patients"],"summary":"Update an existing patient","operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PatientResponseDTO"}}}}}},"delete":{"tags":["Patients"],"summary":"Delete a patient","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/appointments/{id}":{"get":{"tags":["Appointments"],"summary":"Get appointment by ID","operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AppointmentResponseDTO"}}}}}},"put":{"tags":["Appointments"],"summary":"Update an existing appointment","operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppointmentRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AppointmentResponseDTO"}}}}}},"delete":{"tags":["Appointments"],"summary":"Delete an appointment","operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/patients":{"get":{"tags":["Patients"],"summary":"List all patients","operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PatientResponseDTO"}}}}}}},"post":{"tags":["Patients"],"summary":"Create a new patient","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PatientResponseDTO"}}}}}}},"/api/appointments":{"get":{"tags":["Appointments"],"summary":"List all appointments, optionally filtered","operationId":"getAll_1","parameters":[{"name":"patientName","in":"query","required":false,"schema":{"type":"string"}},{"name":"date","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"status","in":"query","required":false,"schema":{"type":"string"}},{"name":"specialty","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AppointmentResponseDTO"}}}}}}},"post":{"tags":["Appointments"],"summary":"Create a new appointment for a patient","operationId":"create_1","parameters":[{"name":"patientId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppointmentRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AppointmentResponseDTO"}}}}}}}},"components":{"schemas":{"PatientRequestDTO":{"type":"object","properties":{"name":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"phoneNumber":{"type":"string"},"email":{"type":"string"}}},"AppointmentSummaryDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"PatientResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"phoneNumber":{"type":"string"},"email":{"type":"string"},"appointments":{"type":"array","items":{"$ref":"#/components/schemas/AppointmentSummaryDTO"}}}},"AppointmentRequestDTO":{"type":"object","properties":{"dateTime":{"type":"string","format":"date-time"},"specialty":{"type":"string"},"status":{"type":"string"}}},"AppointmentResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dateTime":{"type":"string","format":"date-time"},"specialty":{"type":"string"},"status":{"type":"string"},"patientId":{"type":"integer","format":"int64"},"patientName":{"type":"string"}}}}}}